// src/shared/constants/enums.ts
export enum Gender {
  Male,
  Female,
}

export const applyGenderOptions = Object.entries(Gender)
  .filter(([key, value]) => !isNaN(Number(value)))
  .map(([key, value]) => ({
    label: key,
    value: value.toString(),
  }));

export enum EnglishProficiency {
  Excellent,
  Good,
  Average,
  None,
}

export const applyForEnglishProficiencyOptions = Object.keys(EnglishProficiency)
  .filter((key) => isNaN(Number(key)))
  .map((key, value) => ({
    label: key,
    value: value.toString(),
  }));

export enum ApplyForAcademicYear {
  Year2024,
  Year2025,
  Year2026,
}

export const applyForAcademicYearOptions = Object.keys(ApplyForAcademicYear)
  .filter((key) => isNaN(Number(key)))
  .map((key, value) => ({
    label: key.replace("Year", ""),
    value: value.toString(),
  }));

export enum IntakeTerm {
  Semester1,
  Semester2,
  Semester3,
}

export const applyForIntakeTermsOptions = Object.keys(IntakeTerm)
  .filter((key) => isNaN(Number(key)))
  .map((key, value) => ({
    label: key.replace("Semester", "Semester "),
    value: value.toString(),
  }));